+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes qsf-integration project will be documented in this file.
 ## [UNRELEASED]
 ### Fixed
 - fix http oom
+- adjust min range max range for slider in case min < min range or max > max range
 
 ## [5.6]
 ### Added